moved pageSize to pages, for consistency's sake
If pageSize is kept seperately both in the pages and in the repository, there is a risk of inconsistencies, which could cause issues with pagination. pageSize has been moved to the two timeline files. This does mean a bit of code duplication is going on, though the worst inconsistency that could stem from this would be the two timelines showing different amounts of cheeps on a page